Player Info on Title Screen
Be able to see the player Username at the Title Screen. DevyAccount is being worked on. DevyAccount is used for if you have donated to this project or me and you receive something special in return. Either a feature/cosmetics (never) or a way to contact me directly within the client or a exclusive discord server. Yeah that's pretty much it in this commit.